Expanding dragon eye
I put the eye code inside its own component going to use animejs to make the cursor following effect more immersive.
